[test_loop] Refactor all test loop tests into a separate module (#11608)
This PR moves all TestLoop related tests into a separate module in the
integration tests directory

Now all testloop tests and utils live under
`integration-tests/src/test_loop` and are not exposed outside.

This forces all new testloop tests to be a part of this directory else
it would be extremely hard to manage integration tests that are a
mixture of the old style and the new testloop style of tests.

All old set of tests (except a couple of utils) now live under
`integration-tests/src/test`