Unload obsolete memtries post block processing (#11657)
Currently we remove obsolete memtries only when we start state sync for
a new shard:
https://github.com/near/nearcore/blob/master/chain/client/src/client.rs#L2507-L2519
However, it does not unload obsolete memtries in all cases.
For example, if we tracked shards [0, 1, 2], and then we track shards
[0, 1] only, shard 2 won't be unloaded, because we did not start a new
state sync.
Although it is not critical, trie for shard 2 could occupy memory
forever.

This PR calls the routine to unload obsolete memtries as a block
post-processing step too.