fix(nayduck) - Recompile the test contracts when test_features is enabled.  (#11665)
Another attempt at fixing slow_chunk and congestion_control nayduck
tests. Those were failing because the test contract was not being built
correctly with the test_features enabled. Here I'm making the build
script rerun the build if the relevant feature is changed.

The first attempt and some relevant discussion is here - #11619. I know
this new approach is not the best way to fix this issue but it is much
quicker and seems to be getting the job done.

https://nayduck.nearone.org/#/run/160