Change apply-chunk tracker implementation (#11653)
As discussed in [#11447](https://github.com/near/nearcore/issues/11447),
if there is a panic (due to assertion failure) when running testloop
locally, the test freezes instead of reporting the failure. This makes
it difficult to run testloop in cases where we want to debug an
assertion failure.

The reason for the freezing test failures is that the panic causes the
objects be dropped and the drop handler for Chain blocks, waiting
apply-chunks be finished:
https://github.com/near/nearcore/blob/71d35d9921603d8dea60fcc41c0d5478202fc574/chain/chain/src/chain.rs#L295

However that wait does not end, since setter runs after apply-blocks is
finished:
https://github.com/near/nearcore/blob/71d35d9921603d8dea60fcc41c0d5478202fc574/chain/chain/src/chain.rs#L1831

And it does not get a chance to notify the waiters. For example,
commenting out the drop handler for the Chain allows us to see the
assertion failure, as expected.

To handle this for testloop, we change the implementation as follows:
1. We use a condition variable to wait for apply-chunks processing and
setting when it is done (condition variable naturally fits in this
scenario for waiting the operation be done).
2. We introduce a feature to be enabled when running testloop. This
limits the total wait time for each thread to 5 seconds. This is
testing-only and not enabled in production. This is still not ideal and
does not immediately unblock the waiting threads when there is a panic,
but at least puts a limit on the freeze before the assertion failure is
exposed.

Note: We could also use a technique such as `oneshot::Receiver`, in
which case the waiter will just return whenever the sender is dropped,
but it works with `async` computations only, whereas the apply-chunks is
executed in a multi-threaded fashion.