fix: Save genesis bootstrap congestion info into DB and reuse it at node restarts (#11724)
This addresses [#11702](https://github.com/near/nearcore/issues/11702).

The issue happens for forknet and local testing where the network starts
from the genesis block and the genesis protocol version supports
congestion control. In this case, every time the node restarts, it
attempts to re-generate the genesis congestion info. However, this is
done using the state roots at genesis; these state roots are garbage
collected and causes the bootstrapping to fail.

We fix this by storing the genesis congestion info (CongestionInfo for
each shard) in DB. If it exists, the node directly uses it instead of
attempting to re-compute it from state roots.

We store the congestion info in BlockMisc column with key
`GENESIS_CONGESTION_INFO`.

This PR also adds a testloop test to check if the congestion info is
saved (and not cleaned up by GC) and re-enables the previously disabled
nayducks tests due to this problem.