feat: Shadow tracking (#11689)
Part of: https://github.com/near/near-one-project-tracking/issues/65
An option for non-validator node to track shards of given validator.

During stateful -> stateless protocol upgrade a node will track all
shards and will require a lot of RAM. After the migration we can move
the validator key to a new, smaller node, that does not track all
shards.
To make it with minimal downtime, the new node needs to have appropriate
shards in place and memtries loaded in memory, then we hot swap the
validator key without stopping the new node.
But before that happen the new node is not a validator and we need a way
to tell it which validator's shards it should track.