fix: Fix Nayduck test sync_chunks_from_archival  (#11759)
This fix is partial. It mostly fixes the serialization/deserialization
part.

The real issue is that the archival node fails to sync state due to cold
storage not initialized properly. To alleviate this problem, we disable
GC during the test by increasing the number of epochs to keep.

The real fix should be https://github.com/near/nearcore/pull/11752 (or
something that forces archival nodes in the tests to be initialized with
a cold DB).