Configure cold storage for archival nodes in nayduck tests (#11763)
See the explanation here:
https://github.com/near/nearcore/pull/11752#issue-2399165473

If the archival node does not configure the cold storage, garbage
collection of hot data breaks the state sync. In this PR, we configure
the cold storage by default, if the node has `archive` set to true. This
is done as part of the `init_cluster` method.

This fixes the following nayduck tests without having to disable GC:
* `sync_chunks_from_archival.py`
* `state_sync_routed.py`
* `block_sync_archival.py` (there is still an assertion failure at the
end but it is unrelated)

Related issue: #11760