Avoid crashing upon failing to process self chunk endorsements (#11751)
This is a stopgap measure for the problem highlighted in #11750 

There's a corner case related to switching validator key that brings the
node to temporarily fail to apply self chunk endorsement. I think it
might be better to carry on rather than crash since this doesn't break
the chain.

Also fixes the failing test `validator_switch_key`