Feat(eth-implicit-accounts): Stabilize protocol change (#11765)
This PR includes changes to the wallet contract based on the audit
performed by AuditOne. With the audit complete, we are also ready to
stabilize the eth-implicit accounts protocol feature. This is also done
as part of this PR.

This is a relatively large PR. To make things easier for nearcore
reviewers I recommend looking only at the last commit (which does the
protocol stabilization). The various changes to the wallet contract done
in the other commits have been reviewed by the auditors that raised the
issues each commit is addressing.

---------

Co-authored-by: Adam Chudaś <adam@chudas.net>