Avoid sending same chunk endorsement to same block producer (#11805)
Deduplicate validators that appear as the block producer in the next
`NUM_NEXT_BLOCK_PRODUCERS_TO_SEND_CHUNK_ENDORSEMENT` blocks.

While debugging some nayduck tests, I noticed that the validators that
we send endorsements appear in the computed `block_producers` list
multiple times; this would lead to sending duplicate endorsement
messages to the same producers.

Not sure this will be a problem for non-test environments but wanted to
fix it anyways. To fix it, we collect "unique" AccountIds into the
Vector.