feat(forknet) - configure log config only in new_test (#11812)
My goal is to figure out how to set the log level in forknet. Please
consider this PR only as an RFC for what would be the best approach to
solve this problem. I think in the perfect world the initial setup would
happen only once e.g. in `new_test` or in `reset`. It may also be nice
to add a command to adjust the log config but this can always be done
with `run-cmd` so it's low pri.

I also refactored things a bit, sorry for mixing it with the real
change. The real change is moving `configure_log_config` from
`start_neard` to `new_test`, the rest should be a noop.

---------

Co-authored-by: Marcelo Diop-Gonzalez <marcelo827@gmail.com>