Update nayduck test to reduce transaction failure rate (#11819)
In the current version of the `memtrie_disktrie_switch.py`, we create
the accounts using the same validator's signing key. Since the
create-account transactions are issued in parallel, they may be
reordered and the chain can see non-increasing sequence of nonces. This
is seen when running the test multiple times, where it fails when
creating the accounts first.

To avoid this, we use different validators' signing keys to create the
accounts instead of using the same account, so different nonces are
used.