Add/edit some logs to help the block/chunk-tracing work (#11825)
I am working on some analysis code that will take the logs and produce a
Firefox profiler view for some selected logs.

For this, we need some logs named in a specific way and to contain
specific fields (eg. shard-id for chunk-production related events). This
PR makes some changes in existing logs in preparation to capturing these
events with the necessary fields.