runtime: start lifting contract preparation up through tx runtime layers  (#11810)
Best reviewed commit-by-commit.

This ultimately lifts the contract preparation up through a several
function call layers in the transaction runtime up to the layer where
all the currently necessary data are available.

This PR also establishes in broad strokes where the pipelining decisions
will be made (`ApplyProcessingReceiptState`) and makes some minor
changes to the type to have it contain local receipts (in addition to
the previously contained delayed receipts etc) in a queue of sorts which
would allow the pipelining code to look-ahead of the ongoing processing
work and queue-up preparation of the upcoming contracts there.

This work so far is intended to have no functional changes.

Part of #11319