fix: Fix issues with nayduck test split_storage.py (#11831)
This is not a final fix to `split_storage.py`, because there is still
some nondeterminism that makes the test fail. But the changes below are
needed since without them the test consistently fails.

This PR fixes the following:
1) Change `init_cluster` to add a flag to prevent it from initializing
cold-storage for archival node, since the test initializes the
cold-storage after initializing the cluster. The test checks the state
of the archival node before and after it has cold-storage.
2) Update the `DBKind` checks for the hot db, since the hot DB seems to
be using `DBKind::Hot` only, not `DBKind::Archive` anymore.
3) Replace `time.sleep` calls with waiting for a number of blocks to be
more deterministic on the chain progress.