test: Add Nayduck test for replaying chain from genesis using new replay-archival command  (#11858)
In #11851, we added a new command `neard replay-archival` for replaying
blocks from an archival node. The goal is to test the minimal amount of
data to be able to replay the chain blocks from the genesis.

In this PR, we add a Nayduck test (`replay_chain_from_archive.py`) that
will run the chain for a while and then run this new command to replay
the blocks from the genesis block up to the head.

Since the test setup is similar to `memtrie_disktrie_switch.py` (we
create some accounts, deploy test contract for these accounts and then
do transfers, reads, and writes from/to the test contract), we extract
the common code to a new library (`simple_test.py`) and reuse it in the
new test.