[tests] Avoid top-level test contract names (#11867)
As these break when we start using a config that is closer to mainnet
settings in https://github.com/near/nearcore/pull/11864 that prohibits
to create top-level contracts by contracts different from "registrar".

We can alternatively lift that restriction for tests, but I think this
fix is quite simple and aligns the tests with mainnet conventions.